+++ crate/taffy_model/src/taffy_node_mappings.rs
@@ -17,3 +17,21 @@ pub struct TaffyNodeMappings {
     /// `syntect` highlighted spans of entity descriptions.
     pub entity_highlighted_spans: EntityHighlightedSpans,
 }
+
+impl PartialEq for TaffyNodeMappings {
+    fn eq(&self, other: &Self) -> bool {
+        let self_root = self.node_inbuilt_to_taffy.get(&NodeInbuilt::Root).copied();
+        let other_root = other.node_inbuilt_to_taffy.get(&NodeInbuilt::Root).copied();
+        self_root == other_root
+            && self_root
+                .zip(other_root)
+                .map(|(self_root, other_root)| {
+                    self.taffy_tree.children(self_root).ok()
+                        == other.taffy_tree.children(other_root).ok()
+                })
+                .unwrap_or_default()
+            && self.node_inbuilt_to_taffy == other.node_inbuilt_to_taffy
+            && self.node_id_to_taffy == other.node_id_to_taffy
+            && self.entity_highlighted_spans == other.entity_highlighted_spans
+    }
+}